+++ opty/tests/test_direct_collocation.py
@@ -344,8 +344,128 @@ def test_extra_algebraic(plot=False):
     initial_guess = np.zeros(prob.num_free)
     solution, _ = prob.solve(initial_guess)
 
+    with raises(AttributeError):
+        prob.eom_bounds = {0: (-0.5, -0.25),
+                           1: (0.25, 0.5)}
+
     if plot:
         prob.plot_trajectories(solution)
+
+        # Plot the constraint violations. With different eom_bounds, different
+        # settings of the kwargs to plot_constraint_violations.
+        # 2nd test with 5 bounds, balance with 2 bounds on eoms.
+        ax = prob.plot_constraint_violations(solution, show_bounds=True)
+        ax[0].set_title('One plot of all eom violations, '
+                        'no eom_bounds available')
+
+        eom_bounds = {0: (-0.5, 0.5),
+                      1: (-1.0, 1.0),
+                      2: (-2.0, 2.0),
+                      3: (-2.0, 2.0),
+                      4: (-0.1, 0.1)}
+        prob = Problem(
+            obj,
+            obj_grad,
+            eom,
+            states,
+            num_nodes,
+            interval_value,
+            known_parameter_map=par_map,
+            instance_constraints=instance_constraints,
+            eom_bounds=eom_bounds,
+            time_symbol=t,
+            backend='numpy',
+        )
+        ax = prob.plot_constraint_violations(solution, subplots=True,
+                                             show_bounds=True)
+        ax[0].set_title('Separate subplots of eom violations, bounds shown')
+
+        # Balance checks only with two bounds on the eoms
+        # ===============================================
+
+        eom_bounds = {0: (1.0, 2.0),
+                      1: (-1.0, 1.0)}
+        prob = Problem(
+            obj,
+            obj_grad,
+            eom,
+            states,
+            num_nodes,
+            interval_value,
+            known_parameter_map=par_map,
+            instance_constraints=instance_constraints,
+            eom_bounds=eom_bounds,
+            time_symbol=t,
+            backend='numpy',
+        )
+        ax = prob.plot_constraint_violations(solution, subplots=True,
+                                             show_bounds=True)
+        ax[0].set_title('Separate subplots of eom violations, bounds shown, '
+                        ' \n Eq0 violates bounds')
+
+        eom_bounds = {0: (-np.inf, 2.0),
+                      1: (-1.0, np.inf)}
+        prob = Problem(
+            obj,
+            obj_grad,
+            eom,
+            states,
+            num_nodes,
+            interval_value,
+            known_parameter_map=par_map,
+            instance_constraints=instance_constraints,
+            eom_bounds=eom_bounds,
+            time_symbol=t,
+            backend='numpy',
+        )
+        ax = prob.plot_constraint_violations(solution, subplots=True,
+                                             show_bounds=True)
+        ax[0].set_title('Separate subplots of eom violations, bounds shown, '
+                        '\n only finite bounds are shown')
+
+        eom_bounds = {0: (-0.5, 2.0),
+                      1: (0.5, 1.0)}
+        prob = Problem(
+            obj,
+            obj_grad,
+            eom,
+            states,
+            num_nodes,
+            interval_value,
+            known_parameter_map=par_map,
+            instance_constraints=instance_constraints,
+            eom_bounds=eom_bounds,
+            time_symbol=t,
+            backend='numpy',
+        )
+        ax = prob.plot_constraint_violations(solution, subplots=True,
+                                             show_bounds=False)
+        ax[0].set_title('Separate subplots of eom violations, no bounds shown'
+                        '\n Only violations shown')
+
+        eom_bounds = {0: (-0.5, -0.25),
+                      1: (0.25, 0.5)}
+        prob = Problem(
+            obj,
+            obj_grad,
+            eom,
+            states,
+            num_nodes,
+            interval_value,
+            known_parameter_map=par_map,
+            instance_constraints=instance_constraints,
+            eom_bounds=eom_bounds,
+            time_symbol=t,
+            backend='numpy',
+        )
+        ax = prob.plot_constraint_violations(solution, show_bounds=True)
+        ax[0].set_title('Only one plot of eom violations, no bounds shown, \n'
+                        'Violations shown')
+
+        # NOTE :Import is here to avoid a matplotlib import on standard test
+        # runs.
+        import matplotlib.pyplot as plt
+        plt.show()
 
 
 def test_pendulum():